Removed stream/textstream.h from the codebase, not needed anymore.

+++ plugins/eventmanager/Modifiers.cpp
@@ -1,8 +1,7 @@
 #include "Modifiers.h"
 
+#include "itextstream.h"
 #include "iregistry.h"
-
-#include "stream/textstream.h"
 
 #include <boost/lexical_cast.hpp>
 #include <boost/algorithm/string/classification.hpp>
@@ -24,7 +23,7 @@ void Modifiers::loadModifierDefinitions() {
 		
 		if (modifierList.size() > 0) {
 			globalOutputStream() << "EventManager: Modifiers found: " 
-								 << static_cast<int>(modifierList.size()) << "\n";
+								 << modifierList.size() << "\n";
 			for (unsigned int i = 0; i < modifierList.size(); i++) {
 				const std::string name = modifierList[i].getAttributeValue("name");
